requirements.txt
astunparse==1.6.3
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.2.0
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
docopt==0.6.2
dominate==2.8.0
fonttools==4.38.0
idna==3.4
imageio==2.31.1
ipdb==0.13.13
ipython==7.34.0
jedi==0.18.2
jittor==1.3.8.4
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==22.3.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.3.0
requests==2.31.0
scikit-image==0.19.3
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
tifffile==2021.11.2
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.23
wcwidth==0.2.6
wheel==0.38.4
yarg==0.1.9
